Switch to in-house calendar web API
Google Calendar had many shortcomings and limitations. To list a few:

 - severe, long-lasting uncorrected bugs, and lack of maintenance and
   support from Google teams
 - limitations in event recurrence models
 - limitations in access control features
 - bugs and format limitations when trying to integrate into an extended
   workflow

This was a frustrating and disappointing experience to say the
least, especially for a Google product.

Thus, we now switch away our data consumer clients, to our own new web
events API, hosted at calendar.priyom.org, which provides the same data
and for now maintains compatibility with the Google Calendar API formats
and features that we use, to allow for a seamless transition.

Thanks for your service, I guess, Google Calendar.