Remove event caching to localStorage by widget
localStorage is for persistent data storage, and thus is not a suitable
tool for data caching: for example, it provides no mechanism for data
validation, invalidation, expiration, etc...

We'll stop abusing localStorage to reinvent the wheel, and instead rely
on proven built-in browser and HTTP caching control mechanisms. This
was impossible when using Google Calendar as their servers explicitly
disable all data caching with no way to configure it; however with our
new API, this works just fine out of the box once we make event requests
reproducible.